requirements.txt
asgiref==3.9.1
Django==5.2.4
django-jalali==7.4.0
django-jalali-date==2.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
jalali_core==1.0.0
jdatetime==5.2.0
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2
